@@ -8,13 +8,11 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
     "ketcher-core": "workspace:*",
     "ketcher-react": "workspace:*",
     "ketcher-standalone": "workspace:*",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "^5.0.1",
     "typescript": "^4.5.5",
     "web-vitals": "^2.1.4"
@@ -45,6 +43,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
     "serve": "^13.0.2"
   }
 }
