@@ -2,7 +2,9 @@ import 'react-app-polyfill/ie11'
 import 'react-app-polyfill/stable'
 import 'url-search-params-polyfill'
 import './index.css'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from './App'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const container = document.getElementById('root')
+const root = createRoot(container as HTMLElement)
+root.render(<App />)
